+++ BLL/DAL/Suplemento.cs
@@ -27,11 +27,11 @@ public partial class Suplemento
     }
 
 
-    public decimal ID_SUPLEMENTO { get; set; }
+    public int ID_SUPLEMENTO { get; set; }
 
     public int COD_PRODUCTO { get; set; }
 
-    public Nullable<decimal> CANTIDAD_ENTREGAS { get; set; }
+    public Nullable<int> CANTIDAD_ENTREGAS { get; set; }
 
     public Nullable<double> PRECIO { get; set; }
